+++ tests/scalar/test_basic_sympy.py
@@ -20,7 +20,7 @@ def test_SymPyCCode():
     op = SymPyCCode([xs, ys], xs + ys)
     e = op(xt, yt)
     g = theano.gof.FunctionGraph([xt, yt], [e])
-    fn = theano.link.c.cc.CLinker().accept(g).make_function()
+    fn = theano.link.c.basic.CLinker().accept(g).make_function()
     assert fn(1.0, 2.0) == 3.0
