+++ theano/gpuarray/basic_ops.py
@@ -11,7 +11,7 @@ from theano.gof import COp, ParamsType
 from theano.gof.opt import copy_stack_trace
 from theano.gof.utils import MethodNotDefined
 from theano.gradient import grad_undefined
-from theano.link.c.cc import HideC
+from theano.link.c.basic import HideC
 from theano.scalar import bool as bool_t
 from theano.scalar import int32 as int32_t
 from theano.tensor.basic import Alloc, AllocEmpty, Join, Split, alloc_validate_shape
